add parse doc and move expression_test to oth
